Rename and update schema
