+++ bot/src/wrapper/schema.py
@@ -1,36 +1,28 @@
 import datetime
 import decimal
 from dataclasses import dataclass, field
-from typing import Optional, Self, TypeGuard
+from typing import Self
 
-from .api_schema import CompanyPostOutput, CompanyPostIdOutput
-
-
-def is_post_id_output(src: CompanyPostOutput) -> TypeGuard[CompanyPostIdOutput]:
-    return "current_networth" in dir(src) and "is_bankrupt" in dir(src)
+from ._api_schema import CompanyPostIdOutput
 
 
 @dataclass
 class Company:
+    """A dataclass for basic company information from /company endpoint."""
+
     company_name: str
     owner_id: int
     created_date: datetime.datetime
-    current_networth: Optional[decimal.Decimal] = field(default_factory=decimal.Decimal)
-    is_bankrupt: Optional[bool] = False
+    current_networth: decimal.Decimal | None = field(default_factory=decimal.Decimal)
+    is_bankrupt: bool | None = False
 
     @classmethod
-    def from_dict(cls, src: CompanyPostOutput) -> Self:
-        if is_post_id_output(src):
-            return cls(
-                company_name=src["company_name"],
-                owner_id=src["owner_id"],
-                created_date=datetime.datetime.fromisoformat(src["created_date"]),
-                current_networth=decimal.Decimal(src["current_networth"]),
-                is_bankrupt=src["is_bankrupt"],
-            )
-        else:
-            return cls(
-                company_name=src["company_name"],
-                owner_id=src["owner_id"],
-                created_date=datetime.datetime.fromisoformat(src["created_date"]),
-            )
+    def from_dict(cls, src: CompanyPostIdOutput) -> Self:
+        """Convert json from http endpoint to Company object."""
+        return cls(
+            company_name=src["company_name"],
+            owner_id=src["owner_id"],
+            created_date=datetime.datetime.fromisoformat(src["created_date"]),
+            current_networth=decimal.Decimal(src["current_networth"]),
+            is_bankrupt=src["is_bankrupt"],
+        )
